+++ lib/hooks/useAuth.ts
@@ -14,7 +14,6 @@ export const useAuth = () => {
   const [cognitoErrorIsDismissible, setCognitoErrorIsDismissible] = useState(true);
   const [cognitoErrorCallToActionLink, setCognitoErrorCallToActionLink] = useState("");
   const [cognitoErrorCallToActionText, setCognitoErrorCallToActionText] = useState("");
-  const [username, setUsername] = useState("");
 
   const resetCognitoErrorState = () => {
     setCognitoError("");
@@ -24,7 +23,17 @@ export const useAuth = () => {
     setCognitoErrorCallToActionText("");
   };
   const register = async (
-    { username, password, name }: { username: string; password: string; name: string },
+    {
+      username,
+      password,
+      name,
+      confirmationCallback,
+    }: {
+      username: string;
+      password: string;
+      name: string;
+      confirmationCallback: () => void;
+    },
     { setSubmitting }: FormikHelpers<{ username: string; password: string; name: string }>
   ) => {
     resetCognitoErrorState();
@@ -45,9 +54,7 @@ export const useAuth = () => {
           },
           timeout: process.env.NODE_ENV === "production" ? 60000 : 0,
         });
-
-        setSubmitting(false);
-        setUsername(username);
+        confirmationCallback();
       }
     } catch (err) {
       const axiosError = err as AxiosError;
@@ -69,14 +76,22 @@ export const useAuth = () => {
   const confirm = async (
     {
       username,
+      password,
       confirmationCode,
+      setIsAuthorizationError,
+      confirmationCallback,
     }: {
       username: string;
+      password: string;
       confirmationCode: string;
+      setIsAuthorizationError: (s: boolean) => void;
+      confirmationCallback: () => void;
     },
     { setSubmitting, setErrors }: FormikHelpers<{ username: string; confirmationCode: string }>
   ) => {
     resetCognitoErrorState();
+
+    let confirmationSuccess = true;
     try {
       const token = await getCsrfToken();
       if (token) {
@@ -93,15 +108,10 @@ export const useAuth = () => {
           },
           timeout: process.env.NODE_ENV === "production" ? 60000 : 0,
         });
-
-        // Redirect to the Welcome / Account Created page
-
-        await router.push({
-          pathname: "/signup/account-created",
-        });
       }
     } catch (err) {
       const axiosError = err as AxiosError;
+      confirmationSuccess = false;
       logMessage.error(axiosError);
       if (axiosError.response?.data?.message) {
         const errorResponseMessage = axiosError.response.data.message;
@@ -122,6 +132,40 @@ export const useAuth = () => {
       }
     } finally {
       setSubmitting(false);
+    }
+
+    // Automated sign in with provided crednetials
+    if (!confirmationSuccess) return;
+    try {
+      const response = await signIn<"credentials">("credentials", {
+        redirect: false,
+        username,
+        password,
+      });
+
+      if (response?.error) {
+        const responseErrorMessage = response.error;
+        logMessage.error(responseErrorMessage);
+        if (
+          responseErrorMessage.includes("UserNotFoundException") ||
+          responseErrorMessage.includes("NotAuthorizedException")
+        ) {
+          setIsAuthorizationError(true);
+        } else if (responseErrorMessage.includes("GoogleCredentialsExist")) {
+          await router.push("/admin/login");
+        } else {
+          setCognitoError(t("InternalServiceException"));
+        }
+      } else if (response?.ok) {
+        await router.push("/auth/policy?referer=/signup/account-created");
+      }
+    } catch (err) {
+      logMessage.error(err);
+      setCognitoError(t("InternalServiceException"));
+      // Internal error on sign in, not confirmation, so redirect to login page
+      await router.push("/auth/login");
+    } finally {
+      confirmationCallback();
     }
   };
 
@@ -161,7 +205,17 @@ export const useAuth = () => {
   };
 
   const login = async (
-    { username, password }: { username: string; password: string },
+    {
+      username,
+      password,
+      needsConfirmation,
+      didConfirm,
+    }: {
+      username: string;
+      password: string;
+      needsConfirmation: React.Dispatch<React.SetStateAction<boolean>>;
+      didConfirm: boolean;
+    },
     { setSubmitting, setErrors }: FormikHelpers<{ username: string; password: string }>
   ) => {
     resetCognitoErrorState();
@@ -172,13 +226,12 @@ export const useAuth = () => {
         password,
       });
 
-      logMessage.error(response);
       if (response?.error) {
         const responseErrorMessage = response.error;
         setSubmitting(false);
 
         if (responseErrorMessage.includes("UserNotConfirmedException")) {
-          setUsername(username);
+          needsConfirmation(true);
         } else if (
           responseErrorMessage.includes("UserNotFoundException") ||
           responseErrorMessage.includes("NotAuthorizedException")
@@ -194,9 +247,11 @@ export const useAuth = () => {
           setCognitoError(t("InternalServiceException"));
         }
       } else if (response?.ok) {
-        await router.push({
-          pathname: "/myforms",
-        });
+        if (didConfirm) {
+          await router.push("/auth/policy?referer=/signup/account-created");
+        } else {
+          await router.push("/auth/policy");
+        }
       }
     } catch (err) {
       logMessage.error(err);
@@ -218,6 +273,5 @@ export const useAuth = () => {
     cognitoErrorCallToActionText,
     cognitoErrorIsDismissible,
     setCognitoError,
-    username,
   };
 };